core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexus-core"
version = "0.1.0"
description = "Unified behavioral analytics core"
requires-python = ">=3.11"
dependencies = [
    "clickhouse-connect>=0.7.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
    "hypersync>=1.0",
    "python-dotenv>=1.0",
    "eth-abi>=5.0",
    "dagster>=1.7",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4", "mypy>=1.9"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D"]
ignore = ["D100", "D104"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
